TSBM-21 Increased stats in size, because LWIP for example uses uin16_t as size. int8_t seems a bit small to me.
It gave an compile error on conversion

+++ src/system/SystemStats.h
@@ -73,8 +73,8 @@ enum
     kNumEntries
 };
 
-typedef int8_t count_t;
-#define CHIP_SYS_STATS_COUNT_MAX INT8_MAX
+typedef uint16_t count_t;
+#define CHIP_SYS_STATS_COUNT_MAX UINT16_MAX
 
 extern count_t ResourcesInUse[kNumEntries];
 extern count_t HighWatermarks[kNumEntries];
